--- a/lib/saxy/emitter.ex
+++ b/lib/saxy/emitter.ex
@@ -6,18 +6,14 @@ defmodule Saxy.Emitter do
   def emit(event_type, data, %State{user_state: user_state, handler: handler} = state) do
     case do_emit(event_type, data, handler, user_state) do
       {:ok, user_state} ->
-        %{state | user_state: user_state}
+        {:ok, %{state | user_state: user_state}}
 
       {:stop, returning} ->
-        throw({:stop, returning})
+        {:stop, %{state | user_state: returning}}
 
       other ->
-        throw({:error, {:invalid_return, other}})
+        {:error, {event_type, other}}
     end
-  end
-
-  def do_emit(:characters, <<>>, _handler, user_state) do
-    {:ok, user_state}
   end
 
   def do_emit(event_type, data, handler, user_state) when is_atom(handler) do