--- a/lib/saxy/handler_error.ex
+++ b/lib/saxy/handler_error.ex
@@ -1,0 +1,13 @@
+defmodule Saxy.HandlerError do
+  defexception [:reason]
+
+  def message(%__MODULE__{} = exception) do
+    {error_type, term} = exception.reason
+
+    format_message(error_type, term)
+  end
+
+  defp format_message(:bad_return, {event, return}) do
+    "unexpected return #{inspect(return)} in #{inspect(event)} event handler"
+  end
+end