## Speed

Benchmarking a random Reddit RSS feed, the new parser gave 22x better performance of the old parser, and 50% faster than erlsom.

```
Operating System: macOS
CPU Information: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
Number of Available Cores: 4
Available memory: 8 GB
Elixir 1.6.1
Erlang 20.2.2
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
parallel: 1
inputs: none specified
Estimated total run time: 21 s

Name               ips        average  deviation         median         99th %
new saxy        746.19        1.34 ms    ±13.15%        1.28 ms        2.28 ms
erlsom          490.55        2.04 ms    ±16.99%        2.06 ms        2.82 ms
old saxy         33.52       29.83 ms     ±7.71%       30.15 ms       36.67 ms

Comparison:
new saxy        746.19
erlsom          490.55 - 1.52x slower
old saxy         33.52 - 22.26x slower
```

## Better error handling

This parser yields better parsing error.